requirements.txt
streamlit
numpy
tensorflow
scikit-image
matplotlib
pillow